+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 4.4.1
+- Updated CAS [Android](https://github.com/cleveradssolutions/CAS-Android/releases) and [iOS](https://github.com/cleveradssolutions/CAS-iOS/releases) dependencies to 4.4.1
+- Renamed `AdError.rejected` to `AdError.codeRejected` for consistency with other error codes.
+- Migration to the new [Developer Docs](https://docs.page/cleveradssolutions/docs/Flutter)
+
 ## 4.3.0
 ### Major changes
 - Version bumped to 4.3 to align with native SDK versions.